@@ -7,7 +7,7 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.0.0.RELEASE</version>
+        <version>2.6.5</version>
     </parent>
 
     <groupId>com.alexandra</groupId>
@@ -22,14 +22,15 @@
     </modules>
 
     <properties>
-        <spring-boot.version>2.0.0.RELEASE</spring-boot.version>
+        <spring-boot.version>2.6.5</spring-boot.version>
         <io.swagger.version>1.0.34</io.swagger.version>
         <io.springfox.version>2.8.0</io.springfox.version>
-        <cucumber.version>1.2.5</cucumber.version>
-        <h2database.version>1.4.196</h2database.version>
-        <lombok.version>1.18.16</lombok.version>
+        <cucumber.version>7.1.0</cucumber.version>
+        <h2database.version>2.1.210</h2database.version>
+        <lombok.version>1.18.22</lombok.version>
         <oracle.version>11.2.0.3</oracle.version>
-        <flyway.version>4.0</flyway.version>
+        <postgres.version>42.3.1</postgres.version>
+        <flyway.version>8.5.0</flyway.version>
     </properties>
 
     <dependencyManagement>
@@ -66,6 +67,16 @@
                 <artifactId>spring-boot-starter-data-jpa</artifactId>
                 <version>${spring-boot.version}</version>
             </dependency>
+            <dependency>
+                <groupId>org.springframework.boot</groupId>
+                <artifactId>spring-boot-starter-jdbc</artifactId>
+                <version>${spring-boot.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>org.springframework.boot</groupId>
+                <artifactId>spring-boot-starter-validation</artifactId>
+            </dependency>
+
 
             <!--Databases-->
             <dependency>
@@ -74,10 +85,15 @@
                 <version>${h2database.version}</version>
             </dependency>
             <dependency>
+                <groupId>org.postgresql</groupId>
+                <artifactId>postgresql</artifactId>
+                <version>${postgres.version}</version>
+            </dependency>
+           <!-- <dependency>
                 <groupId>oracle</groupId>
                 <artifactId>ojdbc6</artifactId>
                 <version>${oracle.version}</version>
-            </dependency>
+            </dependency>-->
             <dependency>
                 <groupId>org.flywaydb</groupId>
                 <artifactId>flyway-core</artifactId>
@@ -110,17 +126,17 @@
 
             <!--Test-->
             <dependency>
-                <groupId>info.cukes</groupId>
+                <groupId>io.cucumber</groupId>
                 <artifactId>cucumber-java</artifactId>
                 <version>${cucumber.version}</version>
             </dependency>
             <dependency>
-                <groupId>info.cukes</groupId>
+                <groupId>io.cucumber</groupId>
                 <artifactId>cucumber-spring</artifactId>
                 <version>${cucumber.version}</version>
             </dependency>
             <dependency>
-                <groupId>info.cukes</groupId>
+                <groupId>io.cucumber</groupId>
                 <artifactId>cucumber-junit</artifactId>
                 <version>${cucumber.version}</version>
             </dependency>
