@@ -1,7 +1,7 @@
 package com.vet.entity;
 
 import com.vet.DateTimeRange;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import static java.time.LocalDateTime.now;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -12,35 +12,35 @@ public class DateTimeRangeOverlapsTest {
     public void whenOneRangeContainsAnother_thenTheyOverlap(){
         DateTimeRange aRange = new DateTimeRange(now().minusMinutes(5), 60);
         DateTimeRange anotherRange = new DateTimeRange(now(), 30);
-        assertThat(aRange.overlaps(anotherRange));
+        assertThat(aRange.overlaps(anotherRange)).isTrue();
     }
 
     @Test
     public void whenEndRangeIsBetweenStartAndEndAnotherRange_thenTheyOverlap(){
         DateTimeRange aRange = new DateTimeRange(now().minusMinutes(30), 60);
         DateTimeRange anotherRange = new DateTimeRange(now(), 60);
-        assertThat(aRange.overlaps(anotherRange));
+        assertThat(aRange.overlaps(anotherRange)).isTrue();
     }
 
     @Test
     public void whenStartRangeIsBetweenStartAndEndAnotherRange_thenTheyOverlap(){
         DateTimeRange aRange = new DateTimeRange(now().minusMinutes(30), 60);
         DateTimeRange anotherRange = new DateTimeRange(now().minusMinutes(60), 60);
-        assertThat(aRange.overlaps(anotherRange));
+        assertThat(aRange.overlaps(anotherRange)).isTrue();
     }
 
     @Test
     public void whenEndFirstRangeIsBeforeStartSecondRange_thenTheyDoNotOverlap(){
         DateTimeRange aRange = new DateTimeRange(now().minusMinutes(30), 30);
         DateTimeRange anotherRange = new DateTimeRange(now().plusMinutes(30), 30);
-        assertThat(!aRange.overlaps(anotherRange));
+        assertThat(aRange.overlaps(anotherRange)).isFalse();
     }
 
     @Test
     public void whenEndFirstRangeIsImmediatelyBeforeStartSecondRange_thenTheyDoNotOverlap(){
         DateTimeRange aRange = new DateTimeRange(now().minusMinutes(30), 30);
         DateTimeRange anotherRange = new DateTimeRange(now(), 30);
-        assertThat(!aRange.overlaps(anotherRange));
+        assertThat(aRange.overlaps(anotherRange)).isFalse();
     }
 
 }
