@@ -2,10 +2,11 @@ package com.vet.patient;
 
 import lombok.Data;
 
-import javax.persistence.*;
-import javax.persistence.Entity;
+import javax.persistence.Column;
+import javax.persistence.Id;
+import javax.persistence.Table;
 
-@Entity
+//@Entity
 @Table(name = "ANIMAL_TYPES")
 public @Data class AnimalType {
 
